refactor: CommunityDetail - 컴포넌트 분리, 변수명 수정
